LayoutTests/platform/efl/inspector-protocol/debugger/regress-133182-expected.txt

CONSOLE MESSAGE: line 49: [1] Testing statement '({}).a.b.c.d;'
CONSOLE MESSAGE: line 50: [1] Paused and about to step
CONSOLE MESSAGE: line 62: [1] Resumed
CONSOLE MESSAGE: line 54: [1] Paused after stepping
CONSOLE MESSAGE: line 62: [1] Resumed
CONSOLE MESSAGE: line 1: TypeError: undefined is not an object (evaluating '({}).a.b')
CONSOLE MESSAGE: line 49: [2] Testing statement 'exceptionBasic();'
CONSOLE MESSAGE: line 50: [2] Paused and about to step
CONSOLE MESSAGE: line 62: [2] Resumed
CONSOLE MESSAGE: line 54: [2] Paused after stepping
CONSOLE MESSAGE: line 62: [2] Resumed
CONSOLE MESSAGE: line 3: TypeError: undefined is not an object (evaluating '({}).a.b')
CONSOLE MESSAGE: line 49: [3] Testing statement 'exceptionDOM();'
CONSOLE MESSAGE: line 50: [3] Paused and about to step
CONSOLE MESSAGE: line 62: [3] Resumed
CONSOLE MESSAGE: line 54: [3] Paused after stepping
CONSOLE MESSAGE: line 62: [3] Resumed
CONSOLE MESSAGE: line 8: NotFoundError: DOM Exception 8: An attempt was made to reference a Node in a context where it does not exist.
CONSOLE MESSAGE: line 49: [4] Testing statement 'exceptionInHostFunction();'
CONSOLE MESSAGE: line 50: [4] Paused and about to step
CONSOLE MESSAGE: line 62: [4] Resumed
CONSOLE MESSAGE: line 54: [4] Paused after stepping
CONSOLE MESSAGE: line 62: [4] Resumed
CONSOLE MESSAGE: line 24: exception in host function
CONSOLE MESSAGE: line 49: [5] Testing statement 'throwString();'
CONSOLE MESSAGE: line 50: [5] Paused and about to step
CONSOLE MESSAGE: line 62: [5] Resumed
CONSOLE MESSAGE: line 54: [5] Paused after stepping
CONSOLE MESSAGE: line 62: [5] Resumed
CONSOLE MESSAGE: line 13: exception string
CONSOLE MESSAGE: line 49: [6] Testing statement 'throwParam({x:1});'
CONSOLE MESSAGE: line 50: [6] Paused and about to step
CONSOLE MESSAGE: line 62: [6] Resumed
CONSOLE MESSAGE: line 54: [6] Paused after stepping
CONSOLE MESSAGE: line 62: [6] Resumed
CONSOLE MESSAGE: line 18: [object Object]
CONSOLE MESSAGE: line 49: [7] Testing statement 'throwParam(new Error('error message'));'
CONSOLE MESSAGE: line 50: [7] Paused and about to step
CONSOLE MESSAGE: line 62: [7] Resumed
CONSOLE MESSAGE: line 54: [7] Paused after stepping
CONSOLE MESSAGE: line 62: [7] Resumed
CONSOLE MESSAGE: line 18: Error: error message
Regression test for https://bugs.webkit.org/show_bug.cgi?id=133182

Stepping after breaking on uncaught exceptions should not crash

PASS - paused for each uncaught exception
